requirements.txt
annotated-types==0.7.0
anyio==4.12.1
audioop-lts==0.2.2
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.4
cachelib==0.13.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
dnspython==2.8.0
eventlet==0.40.4
firebase_admin==7.1.0
Flask==3.1.2
Flask-Session==0.8.0
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-generativeai==0.7.2
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==21.2.0
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
msgspec==0.20.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
python-dotenv==1.0.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-engineio==4.13.1
python-socketio==5.16.1
requests==2.32.5
rich==14.3.2
simple-websocket==1.1.0
soupsieve==2.8.3
SpeechRecognition==3.14.5
SQLAlchemy==2.0.46
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.5
wikipedia==1.4.0
wsproto==1.3.2
